Remove duplicated CaptureState
